fix(product): cast substring offset to int in product-ID generation (#82)
nextProductID passed an integer offset to SUBSTRING(id FROM $1). Because
SUBSTRING(text FROM ...) is ambiguous between the positional (int) and
regex (text) forms, Postgres inferred $1 as text, so the pgx driver
rejected the integer with 'cannot convert 5 to Text' and every
auto-generated product ID returned HTTP 500. This also caused the admin
UI to fall back to its legacy flat-create path, which stored a price-0
variant (surfacing as $0 in admin and $NaN on the storefront).

Cast $1 to int ($1::int) so the positional form is chosen and the offset
is encoded correctly.

+++ product/pkg/infra/pg/product_store.go
@@ -237,8 +237,12 @@ func brandPrefix(brand string) string {
 func (s *ProductSearchStore) nextProductID(ctx context.Context, brand string) (string, error) {
 	prefix := brandPrefix(brand)
 	var seq int
+	// $1 must be cast to int: SUBSTRING(text FROM ...) is ambiguous between the
+	// positional form (int) and the regex form (text). Without the cast Postgres
+	// infers $1 as text, so the driver rejects the integer offset with
+	// "cannot convert N to Text" and product-ID generation fails.
 	err := s.pool.QueryRow(ctx,
-		`SELECT COALESCE(MAX(CAST(SUBSTRING(id FROM $1) AS INTEGER)), 0)
+		`SELECT COALESCE(MAX(CAST(SUBSTRING(id FROM $1::int) AS INTEGER)), 0)
 		 FROM products WHERE id ~ ('^' || $2 || '-[0-9]+$')`,
 		len(prefix)+2, prefix,
 	).Scan(&seq)
